+++ src/dto/LoSanPham.java
@@ -5,7 +5,7 @@ public class LoSanPham {
     private String maLoSP;
     private String maPN;
     private String maSP;
-    private int soLuong;
+    private Double soLuong;
     private String ngayNhap;
     private String ngaySanXuat;
     private String hanSuDung;
@@ -16,7 +16,7 @@ public class LoSanPham {
     }
 
     // Constructor đầy đủ tham số để đồng bộ với SQL
-    public LoSanPham(String maLoSP, String maPN, String maSP, int soLuong,
+    public LoSanPham(String maLoSP, String maPN, String maSP, Double soLuong,
             String ngayNhap, String ngaySanXuat, String hanSuDung,
             double tongTien, String trangThaiXuLy) {
         this.maLoSP = maLoSP;
@@ -55,11 +55,11 @@ public class LoSanPham {
         this.maSP = maSP;
     }
 
-    public int getSoLuong() {
+    public Double getSoLuong() {
         return soLuong;
     }
 
-    public void setSoLuong(int soLuong) {
+    public void setSoLuong(Double soLuong) {
         this.soLuong = soLuong;
     }
